add stopping background processes

+++ app/src/main/java/ru/net/serbis/launcher/sh/Shell.java
@@ -84,6 +84,9 @@ public class Shell
     
     public void stop(String packageName)
     {
-        command("pkill " + packageName);
+        if (check())
+        {
+            command("pkill " + packageName);
+        }
     }
 }
